+++ tests/src/integration/zerochan-test.cpp
@@ -8,7 +8,7 @@ void ZerochanTest::testHtml()
 	QList<Image*> images = getImages("Zerochan", "www.zerochan.net", "regex", "Touhou", "results.html");
 
 	// Convert results
-	QList<int> ids;
+	QList<qulonglong> ids;
 	for (Image *img : images)
 	{
 		ids.append(img->id());
@@ -16,7 +16,7 @@ void ZerochanTest::testHtml()
 
 	// Check results
 	ids = ids.mid(0, 3);
-	QList<int> expected = QList<int>() << 2034435 << 2034432 << 2034431;
+	QList<qulonglong> expected = QList<qulonglong>() << 2034435 << 2034432 << 2034431;
 	QCOMPARE(images.count(), 20);
 	QCOMPARE(ids, expected);
 }
@@ -26,7 +26,7 @@ void ZerochanTest::testRss()
 	QList<Image*> images = getImages("Zerochan", "www.zerochan.net", "rss", "Touhou", "results.rss");
 
 	// Convert results
-	QList<int> ids;
+	QList<qulonglong> ids;
 	for (Image *img : images)
 	{
 		ids.append(img->id());
@@ -34,7 +34,7 @@ void ZerochanTest::testRss()
 
 	// Check results
 	ids = ids.mid(0, 3);
-	QList<int> expected = QList<int>() << 2034435 << 2034432 << 2034431;
+	QList<qulonglong> expected = QList<qulonglong>() << 2034435 << 2034432 << 2034431;
 	QCOMPARE(images.count(), 20);
 	QCOMPARE(ids, expected);
 }
